Do not crash on devices without a LinkKey
If a bluetooth device does not have a LinkKey, for example if it does
not support pairing, then the current code would crash with an unknown
section error.

This fixes that by handling that case gracefully and showing those
devices in a separate list to make it clear that these are not relevant
for this project.